+++ .dir-locals.el
@@ -1,0 +1,3 @@
+((nil . ((indent-tabs-mode . nil)
+         (require-final-newline . t)))
+ (clojure-mode . ((cljr-favor-prefix-notation . nil))))